Revert "Release 0.1.4 (#19)"
This reverts commit 99dee75edc55ab3006c928a0262e68fa0788d5c4.
