+++ .gitignore
@@ -7,6 +7,7 @@ results
 /mvnw.cmd
 
 #For building dependency
+buildCommand.txt
 setting.xml
 
 *.class
