--- buildCommand.txt
@@ -1,1 +1,0 @@
-./mvnw -s setting.xml clean scala:compile scala:doc-jar deploy
