+++ src/main/java/io/gatling/custom/browser/javaapi/actions/ActionsBase.java
@@ -2,6 +2,7 @@ package io.gatling.custom.browser.javaapi.actions;
 
 
 import com.microsoft.playwright.Page;
+import com.microsoft.playwright.Page.NavigateOptions;
 import io.gatling.custom.browser.model.BrowserSession;
 import io.gatling.javaapi.core.Session;
 import io.gatling.javaapi.core.internal.Expressions;
@@ -11,18 +12,26 @@ import java.util.function.Function;
 
 public class ActionsBase {
 
-    private final io.gatling.custom.browser.actions.BrowserAction wrapped;
+    private final io.gatling.custom.browser.actions.actionsList.BrowserBaseAction wrapped;
 
-    public ActionsBase(io.gatling.custom.browser.actions.BrowserAction actions) {
+    public ActionsBase(io.gatling.custom.browser.actions.actionsList.BrowserBaseAction actions) {
         this.wrapped = actions;
     }
 
-    public BrowserActionOpenBuilder open(String url) {
-        return new BrowserActionOpenBuilder(wrapped.open(Expressions.toStringExpression(url)));
+    public BrowserActionOpen open(String url) {
+        return new BrowserActionOpen(wrapped.open(Expressions.toStringExpression(url)));
     }
 
-    public BrowserActionOpenBuilder open(Function<Session, String> url) {
-        return new BrowserActionOpenBuilder(wrapped.open(Expressions.javaFunctionToExpression(url)));
+    public BrowserActionOpen open(Function<Session, String> url) {
+        return new BrowserActionOpen(wrapped.open(Expressions.javaFunctionToExpression(url)));
+    }
+
+    public BrowserActionOpen open(String url, NavigateOptions options){
+        return new BrowserActionOpen(wrapped.open(Expressions.toStringExpression(url),options));
+    }
+
+    public BrowserActionOpen open(Function<Session, String> url, NavigateOptions options) {
+        return new BrowserActionOpen(wrapped.open(Expressions.javaFunctionToExpression(url),options));
     }
 
     public BrowserActionExecuteFlow executeFlow(BiFunction<Page, BrowserSession, BrowserSession> function) {
