+++ src/main/java/io/gatling/custom/browser/javaapi/actions/BrowserActionOpen.java
@@ -1,15 +1,14 @@
 package io.gatling.custom.browser.javaapi.actions;
 
-
-import io.gatling.custom.browser.actions.actionsList.BrowserActionsExecuteFlowBuilder;
+import io.gatling.custom.browser.actions.actionsList.BrowserActionsOpenBuilder;
 import io.gatling.javaapi.core.ActionBuilder;
 import io.gatling.javaapi.core.ChainBuilder;
 
-public class BrowserActionExecuteFlow implements ActionBuilder {
+public class BrowserActionOpen implements ActionBuilder {
 
-    private final BrowserActionsExecuteFlowBuilder wrapped;
+    private final BrowserActionsOpenBuilder wrapped;
 
-    public BrowserActionExecuteFlow(BrowserActionsExecuteFlowBuilder wrapped) {
+    public BrowserActionOpen(BrowserActionsOpenBuilder wrapped){
         this.wrapped = wrapped;
     }
 
@@ -22,4 +21,5 @@ public class BrowserActionExecuteFlow implements ActionBuilder {
     public ChainBuilder toChainBuilder() {
         return ActionBuilder.super.toChainBuilder();
     }
+
 }
