+++ src/main/java/io/gatling/custom/browser/javaapi/actions/BrowserActionsClearContext.java
@@ -1,21 +1,20 @@
 package io.gatling.custom.browser.javaapi.actions;
 
-
-import io.gatling.custom.browser.actions.actionsList.BrowserActionsExecuteFlowBuilder;
+import io.gatling.custom.browser.actions.actionsList.BrowserClearActionsBuilder;
 import io.gatling.javaapi.core.ActionBuilder;
 import io.gatling.javaapi.core.ChainBuilder;
 
-public class BrowserActionExecuteFlow implements ActionBuilder {
+public class BrowserActionsClearContext implements ActionBuilder {
 
-    private final BrowserActionsExecuteFlowBuilder wrapped;
+    private final BrowserClearActionsBuilder wrapped;
 
-    public BrowserActionExecuteFlow(BrowserActionsExecuteFlowBuilder wrapped) {
+    public BrowserActionsClearContext(BrowserClearActionsBuilder wrapped) {
         this.wrapped = wrapped;
     }
 
     @Override
     public io.gatling.core.action.builder.ActionBuilder asScala() {
-        return this.wrapped;
+        return wrapped;
     }
 
     @Override
