+++ src/main/java/io/gatling/custom/browser/javaapi/protocol/BrowserProtocolBuilderBase.java
@@ -10,7 +10,6 @@ public class BrowserProtocolBuilderBase {
 
     LaunchOptions launchOptions = DefaultProtocolOptions.defaultProtocolOptions() ;
     NewContextOptions contextOptions = DefaultProtocolOptions.defaultContextOptions();
-    Boolean webVitalsEnable = DefaultProtocolOptions.defaultWebVitalsEnable();
 
     public BrowserProtocolBuilderBase() {
         this.wrapped = buildProtocol();
@@ -25,13 +24,9 @@ public class BrowserProtocolBuilderBase {
         this.contextOptions = contextOptions;
         return this;
     }
-    public BrowserProtocolBuilderBase enableUIMetrics(){
-        this.webVitalsEnable = true;
-        return this;
-    }
 
     public BrowserProtocolBuilder buildProtocol(){
-        return new BrowserProtocolBuilder(new io.gatling.custom.browser.protocol.BrowserProtocolBuilder(launchOptions,contextOptions, webVitalsEnable));
+        return new BrowserProtocolBuilder(new io.gatling.custom.browser.protocol.BrowserProtocolBuilder(launchOptions,contextOptions));
     }
 
 }
