+++ src/main/scala/io/gatling/custom/browser/actions/ActionsBase.scala
@@ -11,16 +11,15 @@ import io.gatling.custom.browser.utils.Constants.BROWSER_CONTEXT_KEY
 
 import scala.collection.mutable
 
-trait BrowserActionsBase extends StrictLogging {
+trait ActionsBase extends StrictLogging {
 
   val ctx: ScenarioContext
   private val browserComponent = ctx.protocolComponentsRegistry.components(BrowserProtocol.browserProtocolKey)
   private val browserInstance: Browser = browserComponent.browserInstance
   val browserContextsPool: mutable.Map[Long, BrowserContext] = browserComponent.browserContextsPool
   val contextOptions: Browser.NewContextOptions = browserComponent.contextOptions
-  val enableUIMetrics: Boolean = browserComponent.enableUIMetrics
 
-  protected def getBrowserContextFromSession(session: Session): (Page, Session) = {
+  protected def getBrowserContextFromSession(session: Session): (Page,Session) = {
 
     val userID = session.userId
 
@@ -29,21 +28,21 @@ trait BrowserActionsBase extends StrictLogging {
       var browserContext = browserContextsPool(session.userId)
 
       // Check browser is browser open
-      if (!checkIsBrowserOpen(browserContext)) {
+      if(!checkIsBrowserOpen(browserContext)){
         logger.trace(s"browserContextPool contains context for userID-$userID, but browser was closed, create new instance")
         browserComponent.recreateBrowserInstance()
       }
       // Check page is closed
-      if (!checkIsContextHaveActivePage(browserContext)) {
+      if(!checkIsContextHaveActivePage(browserContext)){
         logger.trace(s"browserContextPool contains context for userID-$userID, but page was closed, create new instance")
         browserContext = browserInstance.newContext(contextOptions)
-        browserContextsPool.put(userID, browserContext)
+        browserContextsPool.put(userID,browserContext)
         val page = browserContext.newPage()
-        return (page, session.set(BROWSER_CONTEXT_KEY, page))
+        return (page , session.set(BROWSER_CONTEXT_KEY,page))
       }
 
       logger.trace(s"browserContextPool contains context for userID-$userID")
-      (session(BROWSER_CONTEXT_KEY).as[Page], session)
+      (session(BROWSER_CONTEXT_KEY).as[Page] , session)
 
     } else {
       logger.trace(s"browserContextPool doesn't contains context for userID-$userID, create new")
