--- src/main/scala/io/gatling/custom/browser/model/PageLoadValidator.scala
@@ -1,9 +1,0 @@
-package io.gatling.custom.browser.model
-
-import com.microsoft.playwright.Page.WaitForFunctionOptions
-
-case class PageLoadValidator(expression: String, arg: Any = null, options: WaitForFunctionOptions = null) {
-  def this(expression: String) = this(expression, null, null)
-
-  def this(expression: String, arg: Any) = this(expression, arg, null)
-}
