+++ src/main/scala/io/gatling/custom/browser/protocol/BrowserComponent.scala
@@ -8,7 +8,7 @@ import io.gatling.custom.browser.utils.Constants.BROWSER_CONTEXT_KEY
 
 import scala.collection.concurrent.TrieMap
 
-case class BrowserComponent(playwright: Playwright, launchOptions: BrowserType.LaunchOptions, contextOptions: Browser.NewContextOptions, enableUIMetrics: Boolean)
+case class BrowserComponent(playwright: Playwright, launchOptions: BrowserType.LaunchOptions, contextOptions: Browser.NewContextOptions)
   extends ProtocolComponents with StrictLogging {
