+++ src/main/scala/io/gatling/custom/browser/protocol/BrowserProtocol.scala
@@ -5,9 +5,8 @@ import com.typesafe.scalalogging.StrictLogging
 import io.gatling.core.CoreComponents
 import io.gatling.core.config.GatlingConfiguration
 import io.gatling.core.protocol.{Protocol, ProtocolKey}
-import io.gatling.custom.browser.stats.UIMetricFileWriter
 
-case class BrowserProtocol(options: BrowserType.LaunchOptions, contextOptions: Browser.NewContextOptions, enableUIMetrics: Boolean) extends Protocol {
+case class BrowserProtocol(options: BrowserType.LaunchOptions, contextOptions: Browser.NewContextOptions) extends Protocol {
   type Component = BrowserComponent
 }
 
@@ -16,16 +15,16 @@ object BrowserProtocol extends StrictLogging {
     override def protocolClass: Class[Protocol] = classOf[BrowserProtocol].asInstanceOf[Class[Protocol]]
 
     override def defaultProtocolValue(configuration: GatlingConfiguration): BrowserProtocol = {
-      BrowserProtocol(DefaultProtocolOptions.defaultProtocolOptions, DefaultProtocolOptions.defaultContextOptions, DefaultProtocolOptions.defaultWebVitalsEnable)
+      BrowserProtocol(DefaultProtocolOptions.defaultProtocolOptions, DefaultProtocolOptions.defaultContextOptions)
     }
 
     override def newComponents(coreComponents: CoreComponents): BrowserProtocol => BrowserComponent = browserProtocol => {
       val playwright: Playwright = Playwright.create()
-      val browserComponent = BrowserComponent(playwright, browserProtocol.options, browserProtocol.contextOptions, browserProtocol.enableUIMetrics)
+      val browserComponent = BrowserComponent(playwright, browserProtocol.options, browserProtocol.contextOptions)
+
       coreComponents.actorSystem.registerOnTermination({
         logger.trace("Received termination signal; closing browser and Playwright instances.")
         browserComponent.browserInstance.close(new Browser.CloseOptions().setReason("Received termination signal from Gatling."))
-        UIMetricFileWriter.closeFile()
         playwright.close()
       })
