+++ src/test/java/ui/performance/simulations/BrowserSimulationsJava.java
@@ -5,7 +5,6 @@ import com.microsoft.playwright.BrowserType.LaunchOptions;
 import com.microsoft.playwright.Page;
 import com.microsoft.playwright.Page.NavigateOptions;
 import com.microsoft.playwright.assertions.PlaywrightAssertions;
-import com.microsoft.playwright.options.WaitUntilState;
 import io.gatling.custom.browser.javaapi.BrowserDsl;
 import io.gatling.custom.browser.model.BrowserSession;
 import io.gatling.javaapi.core.*;
@@ -14,17 +13,16 @@ import org.opentest4j.AssertionFailedError;
 import java.util.Map;
 import java.util.function.BiFunction;
 
+import static com.microsoft.playwright.options.WaitUntilState.NETWORKIDLE;
 import static io.gatling.javaapi.core.CoreDsl.*;
 
-@SuppressWarnings("unused")
 public class BrowserSimulationsJava extends Simulation {
 
     ProtocolBuilder browserProtocol = BrowserDsl
             .gatlingBrowser()
             //// This part of setup block is optional
             .withLaunchOptions(new LaunchOptions().setHeadless(false))
-            .withContextOptions(new NewContextOptions().setViewportSize(1920, 1080).setIsMobile(false))
-            .enableUIMetrics()
+            .withContextOptions(new NewContextOptions().setViewportSize(1920, 1080).setIsMobile(true))
             ////
             .buildProtocol();
 
@@ -47,10 +45,9 @@ public class BrowserSimulationsJava extends Simulation {
     ///// [Example#2] How to set execution timing of action based on your logic
 
     BiFunction<Page, BrowserSession, BrowserSession> exampleFlow2 = (page, browserSession) -> {
-        page.navigate("https://playwright.dev/", new NavigateOptions().setWaitUntil(WaitUntilState.NETWORKIDLE));
+        page.navigate("https://playwright.dev/", new NavigateOptions().setWaitUntil(NETWORKIDLE));
 
         /// How to evaluate JS in playwrights see ===> https://playwright.dev/java/docs/evaluating
-
         Map<String, Double> timing = (Map<String, Double>) page.evaluate("performance.timing");
 
         browserSession.setActionStartTime(timing.get("navigationStart").longValue());
@@ -81,13 +78,6 @@ public class BrowserSimulationsJava extends Simulation {
         return browserSession.updateBrowserSession(session);
     };
 
-    //// [Example#5] How to execute browserSessionFunction
-    BiFunction<Page, BrowserSession, BrowserSession> exampleBrowserSessionFunction = (page, browserSession) -> {
-        Session session = browserSession.getJavaSession().set("pageTitle", page.title());
-
-        return browserSession.updateBrowserSession(session);
-    };
-
 
     ScenarioBuilder mainScenario = scenario("test").repeat(1).on(
             group("flow-a").on(
@@ -95,11 +85,7 @@ public class BrowserSimulationsJava extends Simulation {
                     /*
                      *  You can use EL syntax for action name and url
                      */
-                    exec(BrowserDsl.browserAction("#{actionName}").open("#{url}").withNavigateOptions(new Page.NavigateOptions().setWaitUntil(WaitUntilState.NETWORKIDLE))),
-                    /*
-                     *  You can execute some script that not tracking in report
-                     */
-                    exec(BrowserDsl.browserSessionFunction(exampleBrowserSessionFunction)),
+                    exec(BrowserDsl.browserAction("#{actionName}").open("#{url}")),
                     pause(1, 5),
                     exec(BrowserDsl.browserAction(session -> session.getString("actionName")).open(session -> session.getString("url"))),
                     pause(1, 5),
