+++ src/test/java/ui/performance/simulations/README.md
@@ -1,0 +1,3 @@
+### How to run
+
+`mvn scala:compile gatling:test -Dgatling.simulationClass=ui.performance.simulations.BrowserSimulationsJava`
