+++ src/test/java/ui/performance/simulations/regression/Runner.java
@@ -12,14 +12,11 @@ public class Runner {
     // regression.FailedSimulation
     // regression.ClosedPageRecover
     // regression.Smoke
-    // regression.WebVitalsJava
-    // regression.WebVitalsScala
 
 
     public static void main(String[] args) {
-        String[] runParams = new String[]{
-                GatlingCliOptions.Simulation.shortOption(), "ui.performance.simulations.regression.Smoke",
-                GatlingCliOptions.RunDescription.shortOption(), "Custom descriptions for test-" + System.currentTimeMillis(),
+        String [] runParams = new String[]{
+                GatlingCliOptions.Simulation.shortOption(),"ui.performance.simulations.regression.Smoke",
                 GatlingCliOptions.ResultsFolder.shortOption(), "target/gatling"
         };
