+++ src/test/resources/feeders/page.csv
@@ -1,4 +1,4 @@
 name,link,link2
-home,https://gatling.io/,https://gatling.io/community-vs-enterprise
+home,https://gatling.io/,https://gatling.io/products
 docks,https://docs.gatling.io/,https://docs.gatling.io/reference/script/core/injection/
 community,https://community.gatling.io/,https://community.gatling.io/c/open-source/5
