+++ src/test/resources/logback-test.xml
@@ -9,13 +9,12 @@
 	</appender>
 
 	<appender name="FILE" class="ch.qos.logback.core.FileAppender">
-		<file>target/gatling_log/run.log</file>
+		<file>target/log/tests.log</file>
 		<append>false</append>
 		<encoder>
 			<pattern>%-4relative [%thread] %-5level %logger{35} - %msg%n</pattern>
 		</encoder>
 	</appender>
-
 
 	<!-- uncomment and set to DEBUG to log all browserActions error -->
 	<!-- uncomment and set to TRACE to log all browserActions -->
